Code tidyup
