+++ Assets/Scripts/Commands/BaseCommand.cs
@@ -19,12 +19,6 @@ using UnityEngine;
 
 namespace TiltBrush
 {
-    public enum CommandType
-    {
-        BaseCommand,
-        BrushStrokeCommand,
-        DeleteStrokeCommand
-    }
     /// Commands on the undo stack of SketchMemoryScript
     ///
     /// Subclasses should override the virtual methods to
@@ -32,15 +26,29 @@ namespace TiltBrush
     /// the base class or be sure to recurse into children.
     public class BaseCommand : IDisposable
     {
-        public CommandType m_Type;
-        public Guid m_Guid;
-        public int ChildCount
+        private Guid m_Guid;
+        private BaseCommand m_Parent;
+        protected List<BaseCommand> m_Children;
+
+        public int ChildrenCount
         {
             get { return m_Children.Count; }
         }
-        
-        public List<BaseCommand> m_Children;
-        public BaseCommand m_Parent;
+
+        public Guid Guid
+        {
+            get { return m_Guid; }
+        }
+
+        public List<BaseCommand> Children
+        {
+            get { return m_Children.ToList(); }
+        }
+
+        public Guid ParentGuid
+        {
+            get { return m_Parent != null ? m_Parent.Guid : Guid.Empty; }
+        }
 
         protected static Vector3 GetPositionForCommand(Stroke stroke)
         {
@@ -62,7 +70,6 @@ namespace TiltBrush
         /// The constructor should not mutate sketch state.
         public BaseCommand(BaseCommand parent = null)
         {
-            m_Type = CommandType.BaseCommand;
             m_Guid = Guid.NewGuid();
             m_Children = new List<BaseCommand>();
             if (parent != null)
