--- Assets/Scripts/Multiplayer/Photon/PendingCommand.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6cf03b4030928d04d8af0de680a8e17e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
