+++ Assets/Scripts/Multiplayer/Photon/PhotonPlayerRig.cs
@@ -28,9 +28,6 @@ namespace OpenBrush.Multiplayer
         [SerializeField] private Transform headTransform;
         private PlayerRigData transmitData;
 
-        private Dictionary <Guid, Stroke> m_inProgressStrokes;
-        private List<PendingCommand> m_pendingCommands;
-
         public void TransmitData(PlayerRigData data)
         {
             transmitData = data;
@@ -66,9 +63,6 @@ namespace OpenBrush.Multiplayer
                 transientPointer = PointerManager.m_Instance.CreateRemotePointer();
                 transientPointer.SetBrush(BrushCatalog.m_Instance.DefaultBrush);
                 transientPointer.SetColor(App.BrushColor.CurrentColor);
-
-                m_inProgressStrokes = new Dictionary<Guid, Stroke>();
-                m_pendingCommands = new List<PendingCommand>();
             }
         }
 
@@ -87,9 +81,6 @@ namespace OpenBrush.Multiplayer
 
             else
             {
-                TryProcessCommands();
-
-
                 var toolTR = TrTransform.TR(m_Tool.InterpolationTarget.position, m_Tool.InterpolationTarget.rotation);
                 App.Scene.AsScene[transientPointer.transform] = toolTR;
 
@@ -104,216 +95,5 @@ namespace OpenBrush.Multiplayer
             var remoteTR = TrTransform.TR(m_PlayerHead.InterpolationTarget.position, m_PlayerHead.InterpolationTarget.rotation);
             App.Scene.AsScene[headTransform] = remoteTR;
         }
-
-        private bool CheckifChildStillPending(PendingCommand pending)
-        {   
-            Debug.Log(pending.Command.GetType());
-            if (pending.TotalExpectedChildren == pending.Command.ChildCount)
-            {
-                bool moreChildrenToAssign = false;
-
-                foreach (var childCommand in pending.Command.m_Children)
-                {
-                    // has a child present in the pending queue, check them too
-                    var childPending = m_pendingCommands.FirstOrDefault(x => x.Guid == childCommand.m_Guid);
-
-                    if (!childPending.Guid.Equals(default))
-                    {
-                        var childIsStillPending = CheckifChildStillPending(childPending);
-
-                        if (!childIsStillPending)
-                        {
-                            m_pendingCommands.Remove(childPending);
-                        }
-
-                        moreChildrenToAssign |= childIsStillPending;
-                    }
-                }
-
-                return moreChildrenToAssign;
-            }
-
-            else
-            {
-                return true;
-            }
-        }
-
-        private void InvokePreCommands(PendingCommand pendingCommand)
-        {
-            pendingCommand.PreCommandAction.Invoke();
-
-            foreach (var childCommand in pendingCommand.ChildCommands)
-            {
-                InvokePreCommands(childCommand);
-            }
-        }
-
-        private void TryProcessCommands()
-        {
-            if (m_pendingCommands.Count == 0)
-            {
-                return;
-            }
-            Debug.Log($"Queue length: {m_pendingCommands.Count}");
-
-            var command = m_pendingCommands[0];
-
-            bool stillPending = CheckifChildStillPending(command);
-
-            if (stillPending)
-            {
-                return;
-            }
-
-            // All children present, begin execution
-            
-            m_pendingCommands.RemoveAt(0);
-
-            InvokePreCommands(command);
-
-
-            SketchMemoryScript.m_Instance.PerformAndRecordCommand(command.Command, propegate: false);
-
-            TryProcessCommands();
-        }
-
-        private void AddPendingCommand(Action preAction, Guid commandGuid, Guid parentGuid, BaseCommand command, int childCount)
-        {
-            Debug.Log($"{command.GetType()}, {parentGuid}");
-            PendingCommand pendingCommand = new PendingCommand(commandGuid, command, preAction, childCount);
-
-            if (!parentGuid.Equals(default))
-            {
-                var pendingParent = m_pendingCommands.FirstOrDefault(x => x.Guid == parentGuid);
-                pendingParent.ChildCommands.Add(pendingCommand);
-            }
-
-            m_pendingCommands.Add(pendingCommand);
-        }
-
-        private BaseCommand FindParentCommand(Guid parentGuid)
-        {
-            PendingCommand pendingParent = m_pendingCommands.FirstOrDefault(x => x.Guid == parentGuid);
-
-            if (!parentGuid.Equals(default))
-            {
-                return pendingParent.Command;
-            }
-            return null;
-        }
-
-        public void CreateBrushStroke(Stroke stroke, Guid commandGuid, Guid parentGuid = default, int childCount = 0)
-        {
-            Action preAction = () =>
-            {
-                stroke.m_Type = Stroke.Type.NotCreated;
-                stroke.m_IntendedCanvas = App.Scene.MainCanvas;
-                stroke.Recreate(null, App.Scene.MainCanvas);
-                SketchMemoryScript.m_Instance.MemoryListAdd(stroke);
-            };
-
-            var parentCommand = FindParentCommand(parentGuid);
-
-            var command = new BrushStrokeCommand(stroke, parent: parentCommand);
-
-            AddPendingCommand(preAction, commandGuid, parentGuid, command, childCount);
-        }
-
-#region RPCs
-        [Rpc(InvokeLocal = false)]
-        public void RPC_BaseCommand(Guid commandGuid, Guid parentGuid = default, int childCount = 0)
-        {
-            Debug.Log($"Base command child count: {childCount}");
-            var parentCommand = FindParentCommand(parentGuid);
-            var command = new BaseCommand(parent: parentCommand);
-
-            AddPendingCommand(() => {}, commandGuid, parentGuid, command, childCount);
-        }
-
-        [Rpc(InvokeLocal = false)]
-        public void RPC_BrushStrokeFull(NetworkedStroke strokeData, Guid commandGuid, Guid parentGuid = default, int childCount = 0)
-        {
-            var decode = NetworkedStroke.ToStroke(strokeData);
-
-            CreateBrushStroke(decode, commandGuid, parentGuid, childCount);
-        }
-
-        [Rpc(InvokeLocal = false)]
-        public void RPC_BrushStrokeBegin(Guid id, NetworkedStroke strokeData, int finalLength)
-        {
-            var decode = NetworkedStroke.ToStroke(strokeData);
-
-            decode.m_Type = Stroke.Type.NotCreated;
-            decode.m_IntendedCanvas = App.Scene.MainCanvas;
-            
-            Array.Resize(ref decode.m_ControlPoints, finalLength);
-            Array.Resize(ref decode.m_ControlPointsToDrop, finalLength);
-
-            if(m_inProgressStrokes.ContainsKey(id))
-            {
-                Debug.LogError("Shouldn't be here!");
-                return;
-            }
-
-            m_inProgressStrokes[id] = decode;
-
-        }
-        
-        [Rpc(InvokeLocal = false)]
-        public void RPC_BrushStrokeContinue(Guid id, int offset, NetworkedControlPoint[] controlPoints, bool[] dropPoints)
-        {
-            if(!m_inProgressStrokes.ContainsKey(id))
-            {
-                Debug.LogError("shouldn't be here!");
-                return;
-            }
-
-            var stroke = m_inProgressStrokes[id];
-            
-            for(int i = 0; i < controlPoints.Length; ++i)
-            {
-                stroke.m_ControlPoints[offset + i] = NetworkedControlPoint.ToControlPoint(controlPoints[i]);
-                stroke.m_ControlPointsToDrop[offset + i] = dropPoints[i];
-            }
-        }
-
-        [Rpc(InvokeLocal = false)]
-        public void RPC_BrushStrokeComplete(Guid id, Guid commandGuid, Guid parentGuid = default, int childCount = 0)
-        {
-            if(!m_inProgressStrokes.ContainsKey(id))
-            {
-                Debug.LogError("shouldn't be here!");
-                return;
-            }
-
-            var stroke = m_inProgressStrokes[id];
-
-            CreateBrushStroke(stroke, commandGuid, parentGuid, childCount);
-
-            m_inProgressStrokes.Remove(id);
-        }
-
-        [Rpc(InvokeLocal = false)]
-        public void RPC_DeleteStroke(int seed, Guid commandGuid, Guid parentGuid = default, int childCount = 0)
-        {
-            Debug.Log(seed);
-            var foundStroke = SketchMemoryScript.m_Instance.GetMemoryList.Where(x => x.m_Seed == seed).First();
-
-            if (foundStroke != null)
-            {
-                Debug.Log($"Found seed: {foundStroke.m_Seed}");
-
-                var parentCommand = FindParentCommand(parentGuid);
-                var command = new DeleteStrokeCommand(foundStroke, parent: parentCommand);
-
-                AddPendingCommand(() => {}, commandGuid, parentGuid, command, childCount);
-            }
-            else
-            {
-                Debug.LogError($"couldn't find stroke with seed: {seed}");
-            }
-        }
-#endregion
     }
 }
